Refactor authentication actions and components for improved structure and organization
