+++ src/actions/auth.ts
@@ -3,7 +3,7 @@
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 
-import { createClient } from '../../utils/supabase/server'
+import { createClient } from '../utils/supabase/server'
 
 export async function login(formData: FormData) {
   const supabase = await createClient()
